requirements.txt
keboola.component==1.3.3
keboola.http_client
keboola.utils
https://bitbucket.org/kds_consulting_team/kds-team.processor-csv-to-json/get/0.5.0.zip#egg=csv2json
nested-lookup